+++ pipeline/packager.py
@@ -1,11 +1,13 @@
 from __future__ import unicode_literals
 
+from django.contrib.staticfiles.finders import find
 from django.core.files.base import ContentFile
 from django.utils.encoding import smart_str
 
-from pipeline.conf import settings
 from pipeline.compilers import Compiler
 from pipeline.compressors import Compressor
+from pipeline.conf import settings
+from pipeline.exceptions import PackageNotFound
 from pipeline.glob import glob
 from pipeline.signals import css_compressed, js_compressed
 from pipeline.storage import default_storage
@@ -22,7 +24,7 @@ class Package(object):
             paths = []
             for pattern in self.config.get('source_filenames', []):
                 for path in glob(pattern):
-                    if not path in paths:
+                    if not path in paths and find(path):
                         paths.append(str(path))
             self._sources = paths
         return self._sources
@@ -120,7 +122,3 @@ class Packager(object):
         for name in config:
             packages[name] = Package(config[name])
         return packages
-
-
-class PackageNotFound(Exception):
-    pass