+++ HISTORY.rst
@@ -3,19 +3,36 @@
 History
 =======
 
+1.3.1
+-----
 
-1.1.24
+* Improve exceptions hierarchy.
+* Improve our sub-process calls.
+* Update uglify-js documentation. Thanks to Andrey Antukh.
+
+1.3.0
+-----
+
+* Add support Python 3, with some help from Alan Lu.
+* Add support for Django 1.5.
+* Remove support for Django < 1.4.
+* Drop support for Python < 2.6.
+* Drop support for ``staticfiles`` app, in favor of ``django.contrib.staticfiles``.
+* Drop ``PIPELINE`` settings, in favor of ``DEBUG`` to avoid confusion.
+* Drop support for ``jinja2`` temporarily.
+
+1.2.24
 ------
 
 * Fix yui/yuglify settings overriding each other. Thanks to Fábio Santos.
 
-1.1.23
+1.2.23
 ------
 
 * Separate yuglify compressor from YUI compressor.
 * Improve HTML compression middleware.
 
-1.1.22
+1.2.22
 ------
 
 * Better compressor error messages. Thanks to Steven Cummings.
@@ -23,7 +40,7 @@ History
 * Fix packaging metadata. Thanks to Rui Coelho for noticing it.
 * Add documentation about non-packing storage.
 
-1.1.21
+1.2.21
 ------
 
 * Run stylus even if file is considered outdated.